Guest 5 用户数据 - 收藏App 动态 修复-链路修理 tEST

+++ Common/src/main/java/xyz/spc/common/funcpack/page/PageRequest.java
@@ -14,10 +14,10 @@ public class PageRequest {
     /**
      * 当前页
      */
-    private Long current = 1L;
+    private Integer current = 1;
 
     /**
      * 每页显示条数
      */
-    private Long size = ReqRespCT.DEFAULT_PAGE_SIZE;
+    private Integer size = ReqRespCT.DEFAULT_PAGE_SIZE;
 }
